feat: create sign-in, sign-up endpoints

+++ src/YAEC.Packages/Package.Identity/IdentityOptions.cs
@@ -4,7 +4,7 @@ public class IdentityOptions
 {
     public string TokenSigningKey { get; set; } = null!;
     
-    public double AccessTokenDurationInMinutes { get; set; }
+    public int AccessTokenDurationInMinutes { get; set; }
 
     public int RefreshTokenDurationInDays { get; set; }
